move files to sections folder
