+++ webview-ui/src/components/settings/SettingsView.tsx
@@ -11,12 +11,12 @@ import { CheckCheck, FlaskConical, Info, LucideIcon, Settings, SquareMousePointe
 import { memo, useCallback, useEffect, useRef, useState } from "react"
 import { useEvent } from "react-use"
 import { Tab, TabContent, TabHeader, TabList, TabTrigger } from "../common/Tab"
-import BrowserSettingsSection from "./BrowserSettingsSection"
+import BrowserSettingsSection from "./sections/BrowserSettingsSection"
 import { BrowserSettings } from "@shared/BrowserSettings"
-import FeatureSettingsSection from "./FeatureSettingsSection"
+import FeatureSettingsSection from "./sections/FeatureSettingsSection"
 import Section from "./Section"
 import SectionHeader from "./SectionHeader"
-import TerminalSettingsSection from "./TerminalSettingsSection"
+import TerminalSettingsSection from "./sections/TerminalSettingsSection"
 import ApiConfigurationSection from "./sections/ApiConfigurationSection"
 import GeneralSettingsSection from "./sections/GeneralSettingsSection"
 import { convertApiConfigurationToProtoApiConfiguration } from "@shared/proto-conversions/state/settings-conversion"
