+++ webview-ui/src/components/settings/sections/TerminalSettingsSection.tsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from "react"
 import { VSCodeTextField, VSCodeCheckbox, VSCodeDropdown, VSCodeOption } from "@vscode/webview-ui-toolkit/react"
 import { useExtensionState } from "@/context/ExtensionStateContext"
-import TerminalOutputLineLimitSlider from "./TerminalOutputLineLimitSlider"
-import { StateServiceClient } from "../../services/grpc-client"
+import TerminalOutputLineLimitSlider from "../TerminalOutputLineLimitSlider"
+import { StateServiceClient } from "../../../services/grpc-client"
 import { Int64, Int64Request } from "@shared/proto/common"
-import Section from "./Section"
+import Section from "../Section"
 
 interface TerminalSettingsSectionProps {
 	renderSectionHeader: (tabId: string) => JSX.Element | null
